# packages/api-client/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIS-Leg Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/votaciones": {
      "post": {
        "tags": [
          "votaciones"
        ],
        "summary": "Abrir una votación",
        "description": "Valida el body, abre bajo exclusión y devuelve la entidad normalizada.",
        "operationId": "abrir_votacion_api_v1_votaciones_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/SolicitudVotacionSimple"
                  },
                  {
                    "$ref": "#/components/schemas/SolicitudVotacionEspecial"
                  }
                ],
                "title": "Apertura",
                "discriminator": {
                  "propertyName": "tipo_mayoria",
                  "mapping": {
                    "SIMPLE": "#/components/schemas/SolicitudVotacionSimple",
                    "ESPECIAL": "#/components/schemas/SolicitudVotacionEspecial"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaVotacion"
                }
              }
            }
          },
          "409": {
            "description": "Estado, quórum o votación pendiente impiden la apertura.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Body inválido o tipo descriptivo no permitido por la configuración congelada."
          },
          "503": {
            "description": "No puede garantizarse la auditoría obligatoria.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "500": {
            "description": "Fallo inesperado no clasificado (ERROR_INTERNO).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/votaciones/{id}/finalizacion": {
      "post": {
        "tags": [
          "votaciones"
        ],
        "summary": "Finalizar manualmente una votación como inconclusa",
        "description": "Aplica el motivo validado sobre la votación exacta bajo exclusión.",
        "operationId": "finalizar_votacion_api_v1_votaciones__id__finalizacion_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudFinalizarVotacion"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "409": {
            "description": "Rechazo funcional: ESTADO_INCOMPATIBLE, VOTACION_NO_COINCIDE o VOTACION_NO_EN_CURSO.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "El path o body no cumple el contrato estricto de transporte."
          },
          "503": {
            "description": "No puede garantizarse la auditoría obligatoria.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "500": {
            "description": "Fallo inesperado no clasificado (ERROR_INTERNO).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/votaciones/{id}/desempate": {
      "post": {
        "tags": [
          "votaciones"
        ],
        "summary": "Emitir el desempate presidencial",
        "description": "Desempata la votación exacta con la Presidencia vigente del backend.",
        "operationId": "desempatar_votacion_api_v1_votaciones__id__desempate_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudDesempate"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "409": {
            "description": "Rechazo funcional: ESTADO_INCOMPATIBLE, VOTACION_NO_COINCIDE, VOTACION_NO_EMPATADA o DESEMPATE_YA_EMITIDO.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "El path o body no cumple el contrato estricto de transporte."
          },
          "503": {
            "description": "No puede garantizarse la auditoría obligatoria.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          },
          "500": {
            "description": "Fallo inesperado no clasificado (ERROR_INTERNO).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorRespuesta"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BaseMayoria": {
        "type": "string",
        "enum": [
          "VOTOS_COMPUTABLES",
          "PRESENTES",
          "CUERPO"
        ],
        "title": "BaseMayoria",
        "description": "Representa el denominador conceptual de una regla de mayoría.\n\n``VOTOS_COMPUTABLES`` cuenta positivos y negativos. ``PRESENTES`` es el\ntérmino institucional para quienes emitieron voto ordinario en esa\nvotación, incluidas abstenciones. ``CUERPO`` es el total del padrón\ncongelado. La elección se almacena al abrir y WP-011 la consume al cerrar."
      },
...
      "ErrorRespuesta": {
        "properties": {
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "mensaje": {
            "type": "string",
            "title": "Mensaje"
          }
        },
        "type": "object",
        "required": [
          "codigo",
          "mensaje"
        ],
        "title": "ErrorRespuesta",
        "description": "Cuerpo JSON estable de toda respuesta de error funcional/técnica.\n\n``codigo`` es el identificador estable legible por máquina; ``mensaje``\naporta diagnóstico humano seguro (nunca trazas ni datos internos)."
      },
...
      "EstadoVotacion": {
        "type": "string",
        "enum": [
          "EN_CURSO",
          "CERRADA"
        ],
        "title": "EstadoVotacion",
        "description": "Indica exclusivamente si la recepción todavía admite votos."
      },
...
      "FactorEspecial": {
        "type": "number",
        "maximum": 1.0,
        "exclusiveMinimum": 0.0
      },
      "FactorNumerico": {
        "type": "number"
      },
...
      "NumeroVotacion": {
        "type": "integer",
        "minimum": 1.0
      },
...
      "RespuestaVotacion": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "numero_votacion": {
            "type": "integer",
            "title": "Numero Votacion"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "tema": {
            "type": "string",
            "title": "Tema"
          },
          "tipo_mayoria": {
            "$ref": "#/components/schemas/TipoMayoria"
          },
          "factor": {
            "type": "number",
            "title": "Factor"
          },
          "base": {
            "$ref": "#/components/schemas/BaseMayoria"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoVotacion"
          },
          "fecha_hora_apertura": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Hora Apertura"
          }
        },
        "type": "object",
        "required": [
          "id",
          "numero_votacion",
          "tipo",
          "tema",
          "tipo_mayoria",
          "factor",
          "base",
          "estado",
          "fecha_hora_apertura"
        ],
        "title": "RespuestaVotacion",
        "description": "Representación HTTP mínima y normalizada de la votación creada."
      },
...
      "SolicitudDesempate": {
        "properties": {
          "sentido": {
            "type": "string",
            "enum": [
              "POSITIVO",
              "NEGATIVO"
            ],
            "title": "Sentido"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sentido"
        ],
        "title": "SolicitudDesempate",
        "description": "Body cerrado del único comando presidencial definido por DEC-012.\n\nEl cliente solo elige el sentido. La identidad de Presidencia se obtiene\nde la sesión activa dentro del serializador y no puede inyectarse mediante\ncampos adicionales."
      },
      "SolicitudFinalizarVotacion": {
        "properties": {
          "motivo": {
            "type": "string",
            "title": "Motivo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "motivo"
        ],
        "title": "SolicitudFinalizarVotacion",
        "description": "Body exacto de la finalización manual anticipada de DEC-011.\n\n``strict=True`` evita convertir booleanos, números u otras estructuras en\ntexto. El validador normaliza antes de llegar al dominio y rechaza tanto la\ncadena vacía como la compuesta únicamente por espacios."
      },
...
      "SolicitudVotacionEspecial": {
        "properties": {
          "numero_votacion": {
            "$ref": "#/components/schemas/NumeroVotacion"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "tema": {
            "type": "string",
            "title": "Tema"
          },
          "tipo_mayoria": {
            "type": "string",
            "const": "ESPECIAL",
            "title": "Tipo Mayoria"
          },
          "factor": {
            "$ref": "#/components/schemas/FactorEspecial"
          },
          "base": {
            "$ref": "#/components/schemas/BaseMayoria"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "numero_votacion",
          "tipo",
          "tema",
          "tipo_mayoria",
          "factor",
          "base"
        ],
        "title": "SolicitudVotacionEspecial",
        "description": "Variante ESPECIAL con factor finito y una base canónica obligatoria."
      },
      "SolicitudVotacionSimple": {
        "properties": {
          "numero_votacion": {
            "$ref": "#/components/schemas/NumeroVotacion"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "tema": {
            "type": "string",
            "title": "Tema"
          },
          "tipo_mayoria": {
            "type": "string",
            "const": "SIMPLE",
            "title": "Tipo Mayoria"
          },
          "factor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FactorNumerico"
              },
              {
                "type": "null"
              }
            ]
          },
          "base": {
            "type": "string",
            "const": "VOTOS_COMPUTABLES",
            "title": "Base",
            "default": "VOTOS_COMPUTABLES"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "numero_votacion",
          "tipo",
          "tema",
          "tipo_mayoria"
        ],
        "title": "SolicitudVotacionSimple",
        "description": "Variante SIMPLE con factor/base opcionales y salida normalizada.\n\nUn factor omitido, nulo o numéricamente cero es equivalente. Toda base\ndistinta de ``VOTOS_COMPUTABLES`` queda fuera del propio esquema de esta\nvariante y por eso FastAPI responde 422 antes de invocar el servicio."
      },
...
      "TipoMayoria": {
        "type": "string",
        "enum": [
          "SIMPLE",
          "ESPECIAL"
        ],
        "title": "TipoMayoria",
        "description": "Distingue la regla de decisión declarada por Moderación.\n\nEl tipo es autoritativo: nunca se deduce una mayoría simple o especial a\npartir del factor. Los valores coinciden con el contrato REST de DEC-009."
      },
...
    }
  }
}